fix(capabilities): align Seed/Live contracts and dedupe capability helpers
Behavior fixes:
- Keyset pagination: one shared buildPage with cut-only cursor semantics; exact-multiple result sets no longer emit empty-page cursors
- Live billing applyProviderEvent now scopes its audit event to input.workspaceId like Seed
- requireUnexpired treats unparsable stored expiry as expired instead of throwing

Dedup and structure:
- Split webhook-endpoints into contract / delivery-plan / seed / live modules with direct imports
- assertWithinPlanLimitFor moves the count query beside the plan-limit rule
- planForStripeEvent policy table beside applyProviderEvent; verifyStripeSignature refactored onto shared internal/crypto helpers
- requireMemberRowId in workspace-identity replaces two verbatim copies
- Single wire-projection helpers per module replace hand-assembled literals
